Moved everything into plasma_core
